+++ packages/contexts/src/theme/isEnhancedThemeV1Enabled.ts
@@ -11,8 +11,9 @@ export const isEnhancedThemeV1Enabled = (config?: StoreConfig): boolean => {
         config.checkoutSettings,
         'CHECKOUT-7962.update_font_style_on_checkout_page',
     );
-    const newThemeSettingEnabled =
-        config.checkoutSettings.checkoutUserExperienceSettings.checkoutV2Theme ?? false;
+    const newThemeSettingEnabled = Boolean(
+        config.checkoutSettings.checkoutUserExperienceSettings.enhancedCheckoutThemeV1 ?? false,
+    );
 
     return newThemeSettingEnabled && newThemeExperimentEnabled;
 };