+++ packages/core/src/app/customer/CheckoutButtonContainer.tsx
@@ -1,18 +1,26 @@
 import { type CheckoutSelectors, type CheckoutService } from '@bigcommerce/checkout-sdk';
 import React, { type FunctionComponent, lazy, memo, Suspense } from 'react';
 
-import { type CheckoutContextProps, useLocale } from '@bigcommerce/checkout/contexts';
+import {
+    type CheckoutContextProps,
+    useCapabilities,
+    useLocale,
+} from '@bigcommerce/checkout/contexts';
 import { TranslatedString } from '@bigcommerce/checkout/locale';
-import { WalletButtonsContainerSkeleton } from '@bigcommerce/checkout/ui';
+import { LazyContainer, WalletButtonsContainerSkeleton } from '@bigcommerce/checkout/ui';
 import { isExperimentEnabled } from '@bigcommerce/checkout/utility';
 
 import { withCheckout } from '../checkout';
+import { retry } from '../common/utility';
 
 import { getSupportedMethodIds } from './getSupportedMethods';
 import resolveCheckoutButton from './resolveCheckoutButton';
 
-const CheckoutButtonV1Resolver = lazy(
-    () => import(/* webpackChunkName: "wallet-button-v1-resolver" */ './WalletButtonV1Resolver'),
+const CheckoutButtonV1Resolver = lazy(() =>
+    retry(
+        () =>
+            import(/* webpackChunkName: "wallet-button-v1-resolver" */ './WalletButtonV1Resolver'),
+    ),
 );
 
 interface CheckoutButtonContainerProps {
@@ -46,6 +54,13 @@ const CheckoutButtonContainer: FunctionComponent<
     onWalletButtonClick,
 }) => {
     const { language } = useLocale();
+    const {
+        userJourney: { disableWalletButtons },
+    } = useCapabilities();
+
+    if (disableWalletButtons) {
+        return null;
+    }
 
     try {
         checkEmbeddedSupport(availableMethodIds);
@@ -99,22 +114,24 @@ const CheckoutButtonContainer: FunctionComponent<
                 isPaymentStepActive ? { position: 'absolute', left: '0', top: '-100%' } : undefined
             }
         >
-            <p className="optimizedCheckout-headingSecondary sub-header">
-                <TranslatedString id="remote.start_with_text" />
-            </p>
-            <div className="checkout-buttons-auto-layout">
-                <WalletButtonsContainerSkeleton
-                    buttonsCount={availableMethodIds.length}
-                    isLoading={isLoading}
-                >
-                    <div className="checkoutRemote">{renderButtons()}</div>
-                </WalletButtonsContainerSkeleton>
-            </div>
-            <div className="checkout-separator">
-                <span className="optimizedCheckout-headingSecondary sub-header">
-                    <TranslatedString id="remote.or_text" />
-                </span>
-            </div>
+            <LazyContainer>
+                <p className="optimizedCheckout-headingSecondary sub-header">
+                    <TranslatedString id="remote.start_with_text" />
+                </p>
+                <div className="checkout-buttons-auto-layout">
+                    <WalletButtonsContainerSkeleton
+                        buttonsCount={availableMethodIds.length}
+                        isLoading={isLoading}
+                    >
+                        <div className="checkoutRemote">{renderButtons()}</div>
+                    </WalletButtonsContainerSkeleton>
+                </div>
+                <div className="checkout-separator">
+                    <span className="optimizedCheckout-headingSecondary sub-header">
+                        <TranslatedString id="remote.or_text" />
+                    </span>
+                </div>
+            </LazyContainer>
         </div>
     );
 };